fix: scope venue workspace and privatize Local Support controls
Administrators saw every venue taxonomy term as a managed workspace because
the server bootstrap promoted the entire catalog into managed_venues, and the
public event page carried an organizer-only Find local support action whose
visibility came from broad administrative capability rather than an explicit
organizer relationship.

Restrict the workspace selector to venues with canonical active membership
while leaving administrator claim review on the full catalog, require exact
reciprocal artist roster membership for artist organizer identity, and move
support-state management into the private venue workspace and organizer index
where request absence is the canonical not-seeking state.

Recompose the 1,578-line Venue Settings monolith into focused components built
from existing @extrachill/components primitives, replacing duplicated grid,
stack, and context CSS with package Grid and Panel composition.

Refs #496, #497, #498
