+++ blocks/venue-settings/src/team-tab.js
@@ -1,0 +1,322 @@
+/**
+ * WordPress dependencies
+ */
+import { useState } from '@wordpress/element';
+
+/**
+ * External dependencies
+ */
+import {
+	ActionRow,
+	Badge,
+	FieldGroup,
+	Grid,
+	Panel,
+	PanelHeader,
+} from '@extrachill/components';
+
+/**
+ * Internal dependencies
+ */
+import { errorDetails, runAbility } from './api';
+import { Status } from './status';
+
+export const venueSubscriberCsv = ( subscribers ) =>
+	[
+		[ 'user_id', 'email' ],
+		...subscribers.map( ( subscriber ) => [
+			subscriber.user_id,
+			subscriber.email,
+		] ),
+	]
+		.map( ( row ) =>
+			row
+				.map(
+					( value ) =>
+						`"${ String( value ).replaceAll( '"', '""' ) }"`
+				)
+				.join( ',' )
+		)
+		.join( '\r\n' );
+
+export function TeamTab( {
+	venueId,
+	members,
+	invitations,
+	subscribers,
+	onRefresh,
+} ) {
+	const [ email, setEmail ] = useState( '' );
+	const [ owner, setOwner ] = useState( false );
+	const [ action, setAction ] = useState( null );
+	const mutate = async ( name, input, message ) => {
+		setAction( { tone: 'info', message: 'Working...' } );
+		try {
+			await runAbility( name, input );
+			setAction( { tone: 'success', message } );
+			await onRefresh();
+			return true;
+		} catch ( error ) {
+			const details = errorDetails( error );
+			setAction( {
+				tone: details.status === 409 ? 'warning' : 'error',
+				message:
+					details.status === 409
+						? `${ details.message } Refreshing current team state.`
+						: details.message,
+			} );
+			if ( details.status === 409 ) {
+				await onRefresh();
+			}
+			return false;
+		}
+	};
+	const invite = async ( event ) => {
+		event.preventDefault();
+		const sent = await mutate(
+			'extrachill/create-venue-invitation',
+			{ venue_term_id: venueId, email, is_owner: owner },
+			'Invitation queued.'
+		);
+		if ( sent ) {
+			setEmail( '' );
+			setOwner( false );
+		}
+	};
+	const exportSubscribers = () => {
+		const url = window.URL.createObjectURL(
+			new window.Blob( [ venueSubscriberCsv( subscribers ) ], {
+				type: 'text/csv;charset=utf-8',
+			} )
+		);
+		const link = document.createElement( 'a' );
+		link.href = url;
+		link.download = `venue-${ venueId }-email-subscribers.csv`;
+		link.click();
+		window.URL.revokeObjectURL( url );
+	};
+	return (
+		<Grid minColumnWidth="100%">
+			<Panel>
+				<PanelHeader
+					title="Venue email list"
+					description="Only accounts that explicitly share email access with this venue appear here. Addresses always reflect the current account email."
+				/>
+				{ subscribers.length === 0 ? (
+					<p>
+						No accounts currently share email access with this
+						venue.
+					</p>
+				) : (
+					<>
+						<ul className="ec-venue-settings__records">
+							{ subscribers.map( ( subscriber ) => (
+								<li key={ subscriber.user_id }>
+									<a href={ `mailto:${ subscriber.email }` }>
+										{ subscriber.email }
+									</a>
+								</li>
+							) ) }
+						</ul>
+						<ActionRow>
+							<button
+								type="button"
+								className="button-2"
+								onClick={ exportSubscribers }
+							>
+								Download CSV
+							</button>
+						</ActionRow>
+					</>
+				) }
+			</Panel>
+			<Panel>
+				<PanelHeader
+					title="Invite a teammate"
+					description="Invited accounts remain inactive until the signed email invitation is accepted."
+				/>
+				<form onSubmit={ invite }>
+					<FieldGroup
+						label="Email"
+						htmlFor="venue-invite-email"
+						required
+					>
+						<input
+							id="venue-invite-email"
+							type="email"
+							required
+							value={ email }
+							onChange={ ( event ) =>
+								setEmail( event.target.value )
+							}
+						/>
+					</FieldGroup>
+					<label htmlFor="venue-invite-owner">
+						<input
+							id="venue-invite-owner"
+							type="checkbox"
+							checked={ owner }
+							onChange={ ( event ) =>
+								setOwner( event.target.checked )
+							}
+						/>{ ' ' }
+						Venue owner (can manage team)
+					</label>
+					<ActionRow>
+						<button className="button-1" type="submit">
+							Send invitation
+						</button>
+					</ActionRow>
+				</form>
+				<Status state={ action } />
+			</Panel>
+			<Panel>
+				<PanelHeader
+					title="Memberships"
+					description="Capability access is controlled by WordPress; ownership only controls team administration."
+				/>
+				{ members.length === 0 ? (
+					<p>No membership records found.</p>
+				) : (
+					<ul className="ec-venue-settings__records">
+						{ members.map( ( member ) => (
+							<li key={ member.id }>
+								<div>
+									<strong>User #{ member.user_id }</strong>
+									<div>
+										<Badge>{ member.status }</Badge>{ ' ' }
+										{ member.is_owner && (
+											<Badge>owner</Badge>
+										) }
+									</div>
+								</div>
+								{ member.status === 'active' && (
+									<ActionRow>
+										<button
+											type="button"
+											className="button-2 button-small"
+											onClick={ () =>
+												mutate(
+													'extrachill/update-venue-membership',
+													{
+														venue_term_id: venueId,
+														user_id: member.user_id,
+														is_owner:
+															! member.is_owner,
+														expected_version:
+															member.version,
+													},
+													member.is_owner
+														? 'Owner access removed.'
+														: 'Owner access granted.'
+												)
+											}
+										>
+											{ member.is_owner
+												? 'Make member'
+												: 'Make owner' }
+										</button>
+										<button
+											type="button"
+											className="button-link-delete"
+											onClick={ () =>
+												// eslint-disable-next-line no-alert -- Destructive membership action requires explicit confirmation.
+												window.confirm(
+													'Revoke this venue membership?'
+												) &&
+												mutate(
+													'extrachill/revoke-venue-membership',
+													{
+														venue_term_id: venueId,
+														user_id: member.user_id,
+														expected_version:
+															member.version,
+													},
+													'Membership revoked.'
+												)
+											}
+										>
+											Revoke
+										</button>
+									</ActionRow>
+								) }
+							</li>
+						) ) }
+					</ul>
+				) }
+			</Panel>
+			<Panel>
+				<PanelHeader
+					title="Invitations"
+					description="Delivery status and acceptance lifecycle for this venue."
+				/>
+				{ invitations.length === 0 ? (
+					<p>No invitations found.</p>
+				) : (
+					<ul className="ec-venue-settings__records">
+						{ invitations.map( ( invitation ) => (
+							<li key={ invitation.id }>
+								<div>
+									<strong>
+										User #{ invitation.user_id }
+									</strong>
+									<div>
+										<Badge>{ invitation.status }</Badge>{ ' ' }
+										<Badge>
+											{ invitation.delivery_status }
+										</Badge>{ ' ' }
+										{ invitation.is_owner && (
+											<Badge>owner</Badge>
+										) }
+									</div>
+								</div>
+								{ invitation.status === 'pending' && (
+									<ActionRow>
+										<button
+											type="button"
+											className="button-2 button-small"
+											onClick={ () =>
+												mutate(
+													'extrachill/resend-venue-invitation',
+													{
+														venue_term_id: venueId,
+														invitation_id:
+															invitation.id,
+														expected_version:
+															invitation.version,
+													},
+													'Invitation requeued.'
+												)
+											}
+										>
+											Resend
+										</button>
+										<button
+											type="button"
+											className="button-link-delete"
+											onClick={ () =>
+												mutate(
+													'extrachill/cancel-venue-invitation',
+													{
+														venue_term_id: venueId,
+														invitation_id:
+															invitation.id,
+														expected_version:
+															invitation.version,
+													},
+													'Invitation cancelled.'
+												)
+											}
+										>
+											Cancel
+										</button>
+									</ActionRow>
+								) }
+							</li>
+						) ) }
+					</ul>
+				) }
+			</Panel>
+		</Grid>
+	);
+}
